refactor: centraliza controllers de facebook ads
