+++ backend/ads-service/src/test/java/com/marketinghub/ads/FacebookPageControllerTest.java
@@ -1,5 +1,6 @@
 package com.marketinghub.ads;
 
+import com.marketinghub.facebookads.controller.FacebookPageController;
 import com.marketinghub.repository.jpa.ads.FacebookAccountRepository;
 import com.marketinghub.repository.jpa.ads.FacebookPageRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
