+++ backend/ads-service/src/test/java/com/marketinghub/facebookads/controller/FacebookAdSetControllerTest.java
@@ -1,4 +1,4 @@
-package com.marketinghub.facebookads.web;
+package com.marketinghub.facebookads.controller;
 
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
