+++ docs/registros/experimentos.md
@@ -3695,3 +3695,13 @@
   - frontend/src/pages/pipeline/PipelineCrudPage.tsx
   - frontend/src/api/pipeline/types.ts
   - frontend/src/api/pipeline/usePipelines.ts
+
+## 2026-06-06 — Refatoração dos endpoints de Facebook Ads no backend
+
+- solicitação: concentrar os endpoints que lidam com integração do Facebook Ads em um único pacote e aplicar o padrão atual de arquitetura do backend.
+- causa-raiz: os controllers de contas, páginas, interesses, formulários instantâneos, campanhas, pixels e playbook estavam espalhados entre `com.marketinghub.ads`, `com.marketinghub.facebookads.web` e `com.marketinghub.facebookads.playbook.web`, dificultando auditoria do contrato e evolução modular.
+- correção aplicada: os controllers públicos e internos da integração foram movidos para o pacote único `com.marketinghub.facebookads.controller`, mantendo as rotas legadas para compatibilidade com UI e workers; também foi criado o contrato Swagger dedicado `docs/swagger/facebook-ads-swagger.yaml`.
+- arquivos alterados:
+  - `backend/ads-service/src/main/java/com/marketinghub/facebookads/controller/*Controller.java`
+  - `docs/swagger/facebook-ads-swagger.yaml`
+  - `docs/registros/experimentos.md`
